requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
Werkzeug==2.3.7

Flask
mysql-connector-python
Werkzeug
